Panaderia-repo/backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
python-dotenv==1.1.1
psycopg2-binary